requirements.txt
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
psycopg2==2.7.3.1
PyGithub==1.35
PyJWT==1.5.3
SQLAlchemy==1.1.14
Werkzeug==0.12.2
